requirements_test.txt
pytest<5
pytest-xdist
pytest-cov
pytest-mock
pytest-django
pytest-pythonpath
pycodestyle
pydocstyle
flake8
isort
tox>=3.7
